v1.8.6 #9 Issue 220: Fixed marker/clipboard actions when there's quotes in the name - Note to devs: Probably worth using makeEscapedConfigString in more places! Issue 231: Fixed high CPU on Select only track N actions Issue 232: Fixed folder collapse/uncollaspse/small actions Fixed some potential big endian (PPC) issues with (COMMAND_T).user

+++ Xenakios/TrackTemplateActions.cpp
@@ -68,7 +68,7 @@ void SplitFileNameComponents(string FullFileName,vector<string>& FNComponents)
 void DoOpenTrackTemplate(COMMAND_T* t)
 {
 	char templateFNbeginswith[10];
-	sprintf(templateFNbeginswith, "%02d", t->user);
+	sprintf(templateFNbeginswith, "%02d", (int)t->user);
 	vector<string> blah;
 	string IniFileLoc;
 	IniFileLoc.assign(get_ini_file());
@@ -99,7 +99,7 @@ void DoOpenTrackTemplate(COMMAND_T* t)
 void DoOpenProjectTemplate(COMMAND_T* t)
 {
 	char templateFNbeginswith[10];
-	sprintf(templateFNbeginswith, "%02d", t->user);
+	sprintf(templateFNbeginswith, "%02d", (int)t->user);
 	vector<string> blah;
 	string IniFileLoc;
 	IniFileLoc.assign(get_ini_file());
